+++ CHANGELOG.md
@@ -15,5 +15,7 @@
 * db模块支持显示和格式化显示SQL
 * 调整日志优先级：ConsoleLog优先于JDKLog，Log4j2优先于Log4j
 * db模块的SqlRunner中可自定义Wrapper
+* ExcelReader增加read重载方法（pr#4@Gitee）
 
 ### Bug修复
+* 修复ExcelReader空行导致空指针问题（pr#4@Gitee）